+++ sdmx/reader/csv.py
@@ -10,10 +10,13 @@ from typing import TYPE_CHECKING, Literal, MutableSequence, Optional, Sequence, 
 import sdmx.message
 from sdmx.format import list_media_types
 from sdmx.model import common, v30
+from sdmx.reader import base
 from sdmx.reader.base import BaseReader
 
 if TYPE_CHECKING:
     from typing import TypedDict
+
+    import pandas
 
     from sdmx.model import v21
 
@@ -54,8 +57,7 @@ class Reader(BaseReader):
     suffixes = [".csv"]
 
     #: Handlers for individual fields in a CSV record. This collection has exactly the
-    #: same number of handlers as columns in the `source` passed to
-    #: :meth:`read_message`.
+    #: same number of handlers as columns in the `data` passed to :meth:`convert`.
     handlers: Sequence["Handler"]
 
     _dataflow: Optional["common.BaseDataflow"]
@@ -65,9 +67,12 @@ class Reader(BaseReader):
     def __init__(self):
         self.options = Options()
         self.handlers = []
+        self._dataflow = None
+        self._structure = None
+        self._observations = defaultdict(list)
 
-    def read_message(self, source, structure=None, *, delimiter: str = ",", **kwargs):
-        """Read a message from `source`."""
+    def convert(self, data, structure=None, *, delimiter: str = ",", **kwargs):
+        """Read a message from `data`."""
         self.options.delimiter = delimiter
 
         if isinstance(structure, common.BaseDataflow):
@@ -77,10 +82,8 @@ class Reader(BaseReader):
             self._dataflow = None
             self._structure = structure
 
-        self._observations = defaultdict(list)
-
         # Create a CSV reader
-        lines = source.read().decode().splitlines()
+        lines = data.read().decode().splitlines()
         reader = csv.reader(lines, delimiter=self.options.delimiter)
 
         self.inspect_header(next(reader))
@@ -130,7 +133,7 @@ class Reader(BaseReader):
         Raises
         ------
         ValueError
-            if the source contains malformed SDMX-CSV 2.0.0.
+            if the data contain malformed SDMX-CSV 2.0.0.
         """
         handlers: MutableSequence[Optional["Handler"]] = [
             StoreTarget(allowable={"dataflow", "dataprovision", "datastructure"}),
@@ -207,6 +210,48 @@ class Reader(BaseReader):
 
         self.handlers = tuple(filter(None, handlers))
         assert len(self.handlers) == len(header)
+
+
+class DataFrameConverter(base.Converter):
+    @classmethod
+    def handles(cls, data, kwargs) -> bool:
+        import pandas as pd
+
+        return isinstance(data, pd.DataFrame) and "structure" in kwargs
+
+    def convert(
+        self, data: "pandas.DataFrame", structure=None, **kwargs
+    ) -> "sdmx.message.DataMessage":
+        assert 0 == len(kwargs)
+
+        # TEMPORARY Use a Reader instance
+        r = Reader()
+        r._dataflow = structure
+        r._structure = structure.structure
+        r.inspect_header(data.columns.to_list())
+
+        # Parse remaining rows to observations
+        for _, row in data.iterrows():
+            r.handle_row(row.to_list())
+
+        # Create a data message
+        message = sdmx.message.DataMessage(dataflow=r._dataflow)
+
+        # Create 1 data set for each of the 4 ActionType values
+        ds_kw: "DataSetKwargs" = dict(
+            described_by=r._dataflow, structured_by=r._structure
+        )
+        for (*_, action), obs in r._observations.items():
+            a = common.ActionType[
+                {"A": "append", "D": "delete", "I": "information", "R": "replace"}[
+                    action
+                ]
+            ]
+
+            message.data.append(v30.DataSet(action=a, **ds_kw))
+            message.data[-1].add_obs(obs)
+
+        return message
 
 
 class Handler(ABC):