+++ sdmx/tests/reader/test_base.py
@@ -16,7 +16,7 @@ class TestBaseReader:
                 MediaType("", "xml", "2.1", Flag.data, full="application/foo"),
             ]
 
-            def read_message(self, source, dsd=None):
+            def convert(self, source, dsd=None):
                 pass  # pragma: no cover
 
         return cls
@@ -32,14 +32,16 @@ class TestBaseReader:
             ),
             pytest.raises(ValueError, match="Mismatched structure=FOO, dsd=BAR"),
         ):
-            r.read_message(None, structure=dsd0, dsd=dsd1)
+            r.convert(None, structure=dsd0, dsd=dsd1)
 
     def test_detect(self, MinimalReader):
-        assert False is MinimalReader.detect(b"foo")
+        with pytest.warns(DeprecationWarning, match="use Converter.handles"):
+            assert False is MinimalReader.detect(b"foo")
 
     def test_handles_media_type(self, caplog, MinimalReader):
         """:meth:`.handles_media_type` matches even when params differ, but logs."""
-        assert True is MinimalReader.handles_media_type("application/foo; bar=qux")
+        with pytest.warns(DeprecationWarning, match="use Converter.handles"):
+            assert True is MinimalReader.handles_media_type("application/foo; bar=qux")
         assert (
             "Match application/foo with params {'bar': 'qux'}; "
             "expected {'version': '2.1'}" in caplog.messages