+++ sdmx/tests/reader/test_reader_csv.py
@@ -1,9 +1,13 @@
 from functools import lru_cache
-from io import BytesIO
+from io import BytesIO, StringIO
 from typing import TYPE_CHECKING
 
+import pandas as pd
 import pytest
 
+import sdmx.message
+from sdmx import to_sdmx
+from sdmx.model import common
 from sdmx.reader.csv import Handler, NotHandled, Reader
 
 if TYPE_CHECKING:
@@ -19,6 +23,39 @@ class TestHandler:
         assert "<NotHandled>" == repr(NotHandled())
 
 
+class TestDataFrameConverter:
+    @pytest.fixture
+    def df(self) -> pd.DataFrame:
+        """Equivalent to the content of v21/csv/example-01.csv"""
+        buf = StringIO("""
+STRUCTURE,STRUCTURE_ID,ACTION,DIM_1,DIM_2,DIM_3,OBS_VALUE,ATTR_2,ATTR_3,ATTR_1,UPDATED
+dataflow,ESTAT:NA_MAIN(1.6.0),I,A,B,2014-01,12.4,Y,"Normal, special and other values",N,2021-01-22T13:15:41Z
+dataflow,ESTAT:NA_MAIN(1.6.0),I,A,B,2014-02,10.8,Y,"Normal, special and other values",Y,2021-01-22T13:15:41Z
+""")
+        return pd.read_csv(buf)
+
+    def test_to_sdmx(self, df) -> None:
+        """:class:`.DataFrameConverter` can be used through :func:`.to_sdmx`."""
+        dfd = get_dfd()
+
+        result = to_sdmx(df, structure=dfd)  # Function runs
+
+        assert isinstance(result, sdmx.message.DataMessage)  # Returns a data set
+        assert 1 == len(result.data)  # Message has 1 data set
+        ds = result.data[0]
+
+        assert 2 == len(ds)  # Data set has 2 observations
+        o0 = ds.obs[0]
+
+        assert 12.4 == o0.value  # Observation has an expected value
+        assert (  # Observation has the expected key
+            dfd.structure.make_key(
+                common.Key, dict(DIM_1="A", DIM_2="B", DIM_3="2014-01")
+            )
+            == o0.dimension
+        )
+
+
 class TestReader:
     @pytest.mark.parametrize(
         "mt, expected",
@@ -29,7 +66,8 @@ class TestReader:
         ],
     )
     def test_handles_media_type(self, mt, expected) -> None:
-        assert expected is Reader.handles_media_type(mt)
+        with pytest.warns(DeprecationWarning, match="use Converter.handles"):
+            assert expected is Reader.handles_media_type(mt)
 
     @pytest.mark.parametrize(
         "content, exc_text",
@@ -40,11 +78,12 @@ class TestReader:
     )
     def test_inspect_header0(self, content, exc_text) -> None:
         with pytest.raises(ValueError, match=f"Invalid SDMX-CSV 2.0.0: {exc_text}"):
-            Reader().read_message(BytesIO(content))
+            Reader().convert(BytesIO(content))
 
     @pytest.mark.parametrize("value, expected", [(".csv", True), (".xlsx", False)])
     def test_supports_suffix(self, value, expected) -> None:
-        assert expected is Reader.supports_suffix(value)
+        with pytest.warns(DeprecationWarning, match="use Converter.handles"):
+            assert expected is Reader.supports_suffix(value)
 
 
 @lru_cache
@@ -73,7 +112,7 @@ def get_dfd(n_measure: int = 1) -> "v30.Dataflow":
 
 
 @pytest.mark.parametrize_specimens("path", format="csv")
-def test_read_specimen(path):
+def test_read_specimen(path) -> None:
     """Test that the samples from the SDMX-CSV spec can be read."""
     import sdmx
 
@@ -82,7 +121,7 @@ def test_read_specimen(path):
     else:
         dfd = get_dfd()
 
-    kwargs = dict(structure=dfd)
+    kwargs: dict = dict(structure=dfd)
 
     if path.stem == "example-04":
         kwargs.update(delimiter=";")